requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.14
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycron==3.0.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
redis==4.3.4
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
